refactor: kata php-5kyu RGB To Hex Conversion
